sources/requirements7.txt
alabaster==0.7.12
amqstompclient==1.1.12
arrow==0.13.1
attrs==19.1.0
Babel==2.7.0
cachetools==3.1.1
certifi==2018.8.24
chardet==3.0.4
Click==7.0
cycler==0.10.0
docopt==0.6.2
docutils==0.14
elastic-helper==1.1.4
elasticsearch==7.1.0
Flask==1.0.2
Flask-Cors==3.0.7
gunicorn==19.9.0
humanfriendly==4.18
idna==2.7
imagesize==1.1.0
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
limits==1.3
lxml==4.3.0
MarkupSafe==1.0
matplotlib==3.0.2
numpy==1.15.4
openpyxl==3.0.2
packaging==19.0
pandas==0.23.4
passlib==1.7.1
psycopg2==2.7.7
Pygments==2.4.2
pyparsing==2.3.1
pypng==0.0.19
PyQRCode==1.2.1
python-dateutil==2.7.5
python-docx==0.8.10
python-logstash-async==1.4.1
pytz==2018.7
PyYAML==5.1.1
redis==3.0.1
requests==2.19.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==2.1.2
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.2.17
stomp.py==4.1.21
tzlocal==1.5.1
urllib3==1.23
Werkzeug==0.14.1
xlrd==1.2.0
XlsxWriter==1.1.2
